refactor(redux): migrate layer settings dialog
