+++ src/views/layers/LayerList.jsx
@@ -11,7 +11,7 @@ import ListItemIcon from '@material-ui/core/ListItemIcon';
 import ListItemText from '@material-ui/core/ListItemText';
 import AddCircleOutline from '@material-ui/icons/AddCircleOutline';
 
-import { showLayerSettingsDialog } from '~/reducers/dialogs/layer-settings.js';
+import { showLayerSettingsDialog } from '~/features/map/layer-settings-dialog';
 import { addLayer } from '~/features/map/layers';
 import { selectableListOf } from '~/components/helpers/lists';
 import { labelForLayerType, iconForLayerType } from '~/model/layers';
